# filter-core/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Rule API
    version: 0.0.1
paths:
    /api/v1/core/add_rule:
        post:
            tags:
                - Rule
            operationId: Rule_AddRule
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddRuleReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddRuleResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/core/del_rule:
        post:
            tags:
                - Rule
            operationId: Rule_DelRule
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DelRuleReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DelRuleResp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        AddRuleReq:
            type: object
            properties:
                name:
                    type: string
                dm_type:
                    type: string
                matcher_param_list:
                    type: array
                    items:
                        $ref: '#/components/schemas/MatcherParam'
                action_id_list:
                    type: array
                    items:
                        type: string
        AddRuleResp:
            type: object
            properties: {}
        DelRuleReq:
            type: object
            properties:
                rule_id:
                    type: string
        DelRuleResp:
            type: object
            properties: {}
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        MatcherParam:
            type: object
            properties:
                param:
                    type: string
                base_type:
                    type: string
                match_mode:
                    type: string
                value:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: Rule